+++ src/pages/Details/SliderReview/MoreFrom.js
@@ -4,10 +4,10 @@ import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 
-import ShowCard from '../../components/Card/ShowCard';
+import ShowCard from '../../../components/Card/ShowCard';
 import Above from './Above';
 
-import '../styles/slider.css';
+import './slider.css';
 
 const MoreFrom = () => {
   const slider = useRef();
