+++ src/pages/Details/SliderReview/slider.css
