+++ src/pages/Details/popup.css
